Split macOS shell root views
