+++ clients/apple/AlanNative/MacShellRootView.swift
@@ -26,7 +26,7 @@ struct MacShellRootView: View {
                         isCommandTabPresented = true
                     }
                 }
-                    .frame(width: 286)
+                .frame(width: 286)
 
                 ShellWorkspaceView(host: host)
                     .frame(maxWidth: .infinity, maxHeight: .infinity)
@@ -59,1443 +59,5 @@ struct MacShellRootView: View {
         .animation(.easeOut(duration: 0.18), value: isCommandTabPresented)
         .background(ShellWindowPlacementView(metrics: $windowChromeMetrics))
     }
-}
-
-private struct ShellSidebarView: View {
-    @ObservedObject var host: ShellHostController
-    let chromeMetrics: ShellWindowChromeMetrics
-    let openCommandTab: () -> Void
-    @State private var hoveredTabID: String?
-    @State private var hoveredSpaceID: String?
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 14) {
-            sidebarHeader
-            commandLauncher
-            tabSection
-            spaceDock
-        }
-        .padding(.horizontal, 12)
-        .padding(.top, chromeMetrics.trafficLightsTopInset)
-        .padding(.bottom, 15)
-        .frame(maxHeight: .infinity, alignment: .top)
-        .background {
-            ShellMaterialBackgroundView()
-                .ignoresSafeArea(edges: .top)
-        }
-    }
-
-    private var sidebarHeader: some View {
-        HStack(alignment: .center, spacing: 10) {
-            ZStack {
-                RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                    .fill(ShellPalette.sidebarControlStrong)
-                Text("A")
-                    .font(.system(size: 12.5, weight: .bold, design: .rounded))
-                    .foregroundStyle(ShellPalette.accent)
-            }
-            .frame(width: 28, height: 28)
-
-            VStack(alignment: .leading, spacing: 2) {
-                Text(host.selectedSpace?.title ?? "Alan")
-                    .font(.system(size: 15.5, weight: .semibold))
-                    .foregroundStyle(.primary)
-                    .lineLimit(1)
-                Text(spaceSubtitle)
-                    .font(.system(size: 10.5, weight: .medium))
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-            }
-            .layoutPriority(1)
-
-            Spacer(minLength: 8)
-
-            Menu {
-                Button("New Tab") {
-                    _ = host.openTerminalTab()
-                }
-                Button("Open in Alan") {
-                    _ = host.openAlanTab()
-                }
-            } label: {
-                Image(systemName: "plus")
-                    .font(.system(size: 12, weight: .semibold))
-                    .foregroundStyle(.primary)
-                    .frame(width: 26, height: 26)
-                    .background(
-                        RoundedRectangle(cornerRadius: ShellRadii.control, style: .continuous)
-                            .fill(ShellPalette.sidebarControl)
-                    )
-            }
-            .menuStyle(.borderlessButton)
-            .buttonStyle(.plain)
-            .menuIndicator(.hidden)
-        }
-    }
-
-    private var commandLauncher: some View {
-        Button(action: openCommandTab) {
-            HStack(spacing: 10) {
-                Image(systemName: "magnifyingglass")
-                    .font(.system(size: 11, weight: .semibold))
-                    .foregroundStyle(.secondary)
-                Text("Go to or Command...")
-                    .font(.system(size: 12, weight: .medium))
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-                Spacer(minLength: 0)
-                Text("⌘K")
-                    .font(.system(size: 11, weight: .semibold, design: .monospaced))
-                    .foregroundStyle(.tertiary)
-            }
-            .padding(.horizontal, 11)
-            .padding(.vertical, 10)
-            .background(
-                RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                    .fill(ShellPalette.sidebarControl)
-            )
-        }
-        .buttonStyle(.plain)
-        .keyboardShortcut("k", modifiers: [.command])
-    }
-
-    private var tabSection: some View {
-        VStack(alignment: .leading, spacing: 10) {
-            HStack {
-                Text("Tabs")
-                    .font(.system(size: 11, weight: .semibold))
-                    .textCase(.uppercase)
-                    .foregroundStyle(.secondary)
-                Spacer(minLength: 0)
-                if let selectedSpace = host.selectedSpace {
-                    Text("\(selectedSpace.tabs.count)")
-                        .font(.system(size: 11, weight: .semibold, design: .monospaced))
-                        .foregroundStyle(.secondary)
-                }
-            }
-
-            ScrollView(.vertical, showsIndicators: false) {
-                VStack(alignment: .leading, spacing: 4) {
-                    if let selectedSpace = host.selectedSpace {
-                        ForEach(selectedSpace.tabs) { tab in
-                            tabListRow(for: tab)
-                        }
-                    } else {
-                        ShellEmptyStateRow(
-                            title: "No spaces yet",
-                            detail: "Create a space to start a new stack of terminal tabs."
-                        )
-                    }
-                }
-            }
-        }
-        .frame(maxHeight: .infinity, alignment: .top)
-    }
-
-    private var spaceDock: some View {
-        ShellSidebarSection(title: "Spaces", accessory: "⌥⌘1-9") {
-            HStack(spacing: 10) {
-                ScrollView(.horizontal, showsIndicators: false) {
-                    HStack(spacing: 8) {
-                        ForEach(host.spaces) { space in
-                            Button {
-                                host.select(spaceID: space.spaceID)
-                            } label: {
-                                ShellSpaceRailItem(
-                                    title: space.title,
-                                    symbolName: spaceSymbol(for: space),
-                                    attention: space.attention,
-                                    isSelected: host.selectedSpace?.spaceID == space.spaceID,
-                                    isHovered: hoveredSpaceID == space.spaceID
-                                )
-                            }
-                            .buttonStyle(.plain)
-                            .onHover { isHovering in
-                                hoveredSpaceID = isHovering ? space.spaceID : nil
-                            }
-                        }
-                    }
-                    .padding(.vertical, 2)
-                }
-
-                Menu {
-                    Button("New Space") {
-                        _ = host.createTerminalSpace()
-                    }
-                    Button("New Space with Alan") {
-                        _ = host.createAlanSpace()
-                    }
-                } label: {
-                    Image(systemName: "plus")
-                        .font(.system(size: 12.5, weight: .semibold))
-                        .foregroundStyle(.primary)
-                        .frame(width: 30, height: 30)
-                        .background(
-                            RoundedRectangle(cornerRadius: ShellRadii.control, style: .continuous)
-                                .fill(ShellPalette.sidebarControl)
-                        )
-                }
-                .menuStyle(.borderlessButton)
-                .buttonStyle(.plain)
-                .menuIndicator(.hidden)
-                .help("Create a new space")
-            }
-        }
-    }
-
-    private var spaceSubtitle: String {
-        guard let selectedSpace = host.selectedSpace else {
-            return "Terminal-first on macOS"
-        }
-
-        let count = selectedSpace.tabs.count
-        return count == 1 ? "1 tab" : "\(count) tabs"
-    }
-
-    private func close(tab: ShellTab) {
-        host.select(tabID: tab.tabID)
-        _ = host.closeSelectedTab()
-    }
-
-    @ViewBuilder
-    private func tabListRow(for tab: ShellTab) -> some View {
-        let isSelected = host.selectedTab?.tabID == tab.tabID
-        let isHovered = hoveredTabID == tab.tabID
-
-        ShellTabSidebarRow(
-            title: tabTitle(for: tab),
-            subtitle: tabSubtitle(for: tab),
-            iconName: tabIconName(for: tab),
-            attention: strongestAttention(for: tab),
-            showsAlanMarker: showsAlanMarker(for: tab),
-            isSelected: isSelected,
-            isHovered: isHovered,
-            showsMenuAffordance: isHovered,
-            onClose: { close(tab: tab) }
-        )
-        .contentShape(Rectangle())
-        .onTapGesture {
-            host.select(tabID: tab.tabID)
-        }
-        .onHover { isHovering in
-            hoveredTabID = isHovering ? tab.tabID : nil
-        }
-        .contextMenu {
-            Button("New Tab") {
-                _ = host.openTerminalTab()
-            }
-            Button("Open in Alan") {
-                _ = host.openAlanTab()
-            }
-            Divider()
-            Button("Close Tab", role: .destructive) {
-                close(tab: tab)
-            }
-        }
-    }
-
-    private func fallbackTitle(for tab: ShellTab) -> String {
-        switch tab.kind {
-        case .terminal:
-            return "Terminal"
-        case .scratch:
-            return "Scratch"
-        case .log:
-            return "Logs"
-        }
-    }
-
-    private func tabIconName(for tab: ShellTab) -> String {
-        switch tab.kind {
-        case .terminal:
-            return "terminal"
-        case .scratch:
-            return "note.text"
-        case .log:
-            return "doc.text.magnifyingglass"
-        }
-    }
-
-    private func tabTitle(for tab: ShellTab) -> String {
-        let panes = host.shellState.panes.filter { $0.tabID == tab.tabID }
-        let primaryPane = panes.first
-        return shellDisplayTitle(
-            rawTitle: tab.title ?? primaryPane?.viewport?.title,
-            workingDirectoryName: primaryPane?.context?.workingDirectoryName,
-            cwd: primaryPane?.cwd,
-            program: primaryPane?.process?.program,
-            launchTarget: primaryPane?.resolvedLaunchTarget ?? .shell,
-            fallback: fallbackTitle(for: tab)
-        )
-    }
-
-    private func tabSubtitle(for tab: ShellTab) -> String {
-        let panes = host.shellState.panes.filter { $0.tabID == tab.tabID }
-        let primaryPane = panes.first
-        let title = tabTitle(for: tab)
-
-        if let primaryPane,
-           let status = shellTerminalStatusSummary(for: primaryPane)
-        {
-            return status
-        }
-
-        if let branch = primaryPane?.context?.gitBranch,
-           let folder = primaryPane?.context?.workingDirectoryName
-        {
-            if folder == title {
-                return branch
-            }
-            return "\(folder)  ·  \(branch)"
-        }
-
-        if let folder = shellVisibleLabel(primaryPane?.context?.workingDirectoryName) ?? shellPathLeaf(primaryPane?.cwd) {
-            if folder == title, let program = shellVisibleLabel(primaryPane?.process?.program) {
-                return program
-            }
-            return folder
-        }
-
-        if let program = primaryPane?.process?.program {
-            return program
-        }
-
-        return tab.kind.rawValue.capitalized
-    }
-
-    private func strongestAttention(for tab: ShellTab) -> ShellAttentionState? {
-        host.shellState.panes
-            .filter { $0.tabID == tab.tabID }
-            .map(\.attention)
-            .sorted { attentionRank(for: $0) > attentionRank(for: $1) }
-            .first(where: { $0 != .idle })
-    }
-
-    private func spaceSymbol(for space: ShellSpace) -> String {
-        let hasAlan = host.shellState.panes.contains { pane in
-            pane.spaceID == space.spaceID && pane.resolvedLaunchTarget == .alan
-        }
-
-        if hasAlan {
-            return "sparkles"
-        }
-
-        if space.tabs.count > 1 {
-            return "square.stack.3d.up"
-        }
-
-        return "terminal"
-    }
-
-    private func showsAlanMarker(for tab: ShellTab) -> Bool {
-        host.shellState.panes.contains { pane in
-            pane.tabID == tab.tabID && pane.resolvedLaunchTarget == .alan
-        }
-    }
-
-    private func attentionRank(for attention: ShellAttentionState) -> Int {
-        switch attention {
-        case .awaitingUser:
-            return 3
-        case .notable:
-            return 2
-        case .active:
-            return 1
-        case .idle:
-            return 0
-        }
-    }
-}
-
-private struct ShellWorkspaceView: View {
-    @ObservedObject var host: ShellHostController
-
-    var body: some View {
-        TerminalPaneView(host: host)
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-}
-
-private struct ShellSpaceKeyboardShortcuts: View {
-    @ObservedObject var host: ShellHostController
-
-    var body: some View {
-        VStack(spacing: 0) {
-            Button("") {
-                host.selectAdjacentSpace(offset: -1)
-            }
-            .keyboardShortcut(.leftArrow, modifiers: [.command, .option])
-
-            Button("") {
-                host.selectAdjacentSpace(offset: 1)
-            }
-            .keyboardShortcut(.rightArrow, modifiers: [.command, .option])
-
-            ForEach(Array(host.spaces.prefix(9).enumerated()), id: \.element.spaceID) { index, _ in
-                Button("") {
-                    host.selectSpace(at: index)
-                }
-                .keyboardShortcut(
-                    KeyEquivalent(Character(String(index + 1))),
-                    modifiers: [.command, .option]
-                )
-            }
-        }
-        .labelsHidden()
-        .buttonStyle(.plain)
-        .frame(width: 0, height: 0)
-        .opacity(0.001)
-        .allowsHitTesting(false)
-        .accessibilityHidden(true)
-    }
-}
-
-private struct ShellSpaceRailItem: View {
-    @Environment(\.accessibilityReduceMotion) private var reduceMotion
-    let title: String
-    let symbolName: String
-    let attention: ShellAttentionState
-    let isSelected: Bool
-    let isHovered: Bool
-
-    var body: some View {
-        ZStack(alignment: .topTrailing) {
-            ZStack {
-                RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                    .fill(
-                        isSelected
-                            ? ShellPalette.railSelection
-                            : (isHovered ? ShellPalette.railHover : ShellPalette.railBase)
-                    )
-                Image(systemName: symbolName)
-                    .font(.system(size: 11, weight: .semibold))
-                    .foregroundStyle(isSelected ? ShellPalette.accent : .primary)
-            }
-            .frame(width: 34, height: 34)
-
-            if attention != .idle {
-                Circle()
-                    .fill(attentionColor)
-                    .frame(width: 9, height: 9)
-                    .overlay {
-                        Circle()
-                            .stroke(ShellPalette.sidebarCard, lineWidth: 1.5)
-                    }
-                    .offset(x: 2, y: -2)
-            }
-        }
-        .scaleEffect(isSelected ? 1 : (isHovered ? 1.03 : 1))
-        .shadow(color: isSelected ? ShellPalette.accent.opacity(0.12) : .clear, radius: 8, y: 3)
-        .animation(reduceMotion ? nil : .easeOut(duration: 0.16), value: isHovered)
-        .animation(reduceMotion ? nil : .easeOut(duration: 0.16), value: isSelected)
-        .help(title)
-    }
-
-    private var attentionColor: Color {
-        switch attention {
-        case .idle:
-            return ShellPalette.line
-        case .active:
-            return ShellPalette.accent
-        case .awaitingUser:
-            return ShellPalette.attention
-        case .notable:
-            return Color(red: 0.71, green: 0.58, blue: 0.28)
-        }
-    }
-}
-
-private struct ShellTabSidebarRow: View {
-    @Environment(\.accessibilityReduceMotion) private var reduceMotion
-    let title: String
-    let subtitle: String
-    let iconName: String
-    let attention: ShellAttentionState?
-    let showsAlanMarker: Bool
-    let isSelected: Bool
-    let isHovered: Bool
-    let showsMenuAffordance: Bool
-    let onClose: () -> Void
-
-    var body: some View {
-        HStack(spacing: 10) {
-            Image(systemName: iconName)
-                .font(.system(size: 11.5, weight: .semibold))
-                .foregroundStyle(isSelected ? ShellPalette.accent : .secondary)
-                .frame(width: 14)
-
-            VStack(alignment: .leading, spacing: 3) {
-                HStack(spacing: 6) {
-                    Text(title)
-                        .font(.system(size: 13, weight: .semibold))
-                        .tracking(-0.1)
-                        .foregroundStyle(.primary)
-                        .lineLimit(1)
-
-                    if showsAlanMarker {
-                        Image(systemName: "sparkles")
-                            .font(.system(size: 9, weight: .bold))
-                            .foregroundStyle(ShellPalette.accent)
-                    }
-                }
-
-                Text(subtitle)
-                    .font(.system(size: 11, weight: .medium))
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-            }
-
-            Spacer(minLength: 8)
-
-            if let attention {
-                Circle()
-                    .fill(attentionColor(for: attention))
-                    .frame(width: 8, height: 8)
-            }
-
-            if showsMenuAffordance {
-                Button(action: onClose) {
-                    Image(systemName: "xmark")
-                        .font(.system(size: 9, weight: .bold))
-                        .foregroundStyle(.secondary)
-                        .frame(width: 18, height: 18)
-                }
-                .buttonStyle(.plain)
-            }
-        }
-        .padding(.horizontal, 10)
-        .padding(.vertical, 7)
-        .background(
-                RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                    .fill(
-                        isSelected
-                            ? ShellPalette.sidebarSelection
-                            : (isHovered ? ShellPalette.sidebarHover : Color.clear)
-                    )
-            )
-        .scaleEffect(isHovered && !isSelected ? 1.005 : 1)
-        .animation(reduceMotion ? nil : .easeOut(duration: 0.16), value: isHovered)
-        .animation(reduceMotion ? nil : .easeOut(duration: 0.16), value: isSelected)
-    }
-
-    private func attentionColor(for attention: ShellAttentionState) -> Color {
-        switch attention {
-        case .idle:
-            return ShellPalette.line
-        case .active:
-            return ShellPalette.accent
-        case .awaitingUser:
-            return ShellPalette.attention
-        case .notable:
-            return Color(red: 0.71, green: 0.58, blue: 0.28)
-        }
-    }
-}
-
-private struct ShellTabRailView: View {
-    @ObservedObject var host: ShellHostController
-
-    var body: some View {
-        HStack(spacing: 10) {
-            ForEach(host.selectedSpace?.tabs ?? []) { tab in
-                Button {
-                    host.select(tabID: tab.tabID)
-                } label: {
-                    VStack(alignment: .leading, spacing: 5) {
-                        Text(tab.title ?? tab.kind.rawValue.capitalized)
-                            .font(.system(size: 13, weight: .semibold, design: .rounded))
-                        Text(tab.kind.rawValue)
-                            .font(.system(size: 11, weight: .medium, design: .rounded))
-                            .textCase(.uppercase)
-                            .foregroundStyle(host.selectedTab?.tabID == tab.tabID ? ShellPalette.accent.opacity(0.82) : ShellPalette.mutedInk)
-                    }
-                    .foregroundStyle(host.selectedTab?.tabID == tab.tabID ? ShellPalette.ink : ShellPalette.mutedInk)
-                    .padding(.horizontal, 14)
-                    .padding(.vertical, 12)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .background(
-                        RoundedRectangle(cornerRadius: ShellRadii.surface, style: .continuous)
-                            .fill(host.selectedTab?.tabID == tab.tabID ? ShellPalette.sidebarCard : ShellPalette.sidebarHover)
-                    )
-                    .overlay {
-                        RoundedRectangle(cornerRadius: ShellRadii.surface, style: .continuous)
-                            .stroke(ShellPalette.line.opacity(host.selectedTab?.tabID == tab.tabID ? 0.55 : 0.24), lineWidth: 1)
-                    }
-                }
-                .buttonStyle(.plain)
-            }
-        }
-    }
-}
-
-private struct ShellSidebarSection<Content: View>: View {
-    let title: String
-    let accessory: String?
-    @ViewBuilder let content: Content
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack {
-                Text(title)
-                    .font(.system(size: 11, weight: .semibold))
-                    .textCase(.uppercase)
-                    .foregroundStyle(.secondary)
-                Spacer(minLength: 0)
-                if let accessory {
-                    Text(accessory)
-                        .font(.system(size: 11, weight: .semibold, design: .monospaced))
-                        .foregroundStyle(.tertiary)
-                }
-            }
-
-            content
-        }
-    }
-}
-
-private struct ShellSpaceRow: View {
-    let space: ShellSpace
-    let isSelected: Bool
-
-    var body: some View {
-        HStack(alignment: .center, spacing: 10) {
-            Circle()
-                .fill(color(for: space.attention))
-                .frame(width: 8, height: 8)
-
-            VStack(alignment: .leading, spacing: 4) {
-                Text(space.title)
-                    .font(.system(size: 13, weight: .semibold))
-                    .foregroundStyle(ShellPalette.ink)
-                Text("\(space.tabs.count) tab\(space.tabs.count == 1 ? "" : "s")")
-                    .font(.system(size: 11, weight: .medium))
-                    .foregroundStyle(ShellPalette.mutedInk)
-            }
-
-            Spacer(minLength: 0)
-        }
-        .padding(.horizontal, 12)
-        .padding(.vertical, 10)
-        .background(
-            RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                .fill(isSelected ? ShellPalette.sidebarCard : Color.clear)
-        )
-        .overlay {
-            RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                .stroke(isSelected ? ShellPalette.line.opacity(0.6) : Color.clear, lineWidth: 1)
-        }
-    }
-
-    private func color(for attention: ShellAttentionState) -> Color {
-        switch attention {
-        case .idle:
-            return ShellPalette.line
-        case .active:
-            return ShellPalette.accent
-        case .awaitingUser:
-            return ShellPalette.attention
-        case .notable:
-            return Color(red: 0.71, green: 0.58, blue: 0.28)
-        }
-    }
-}
-
-private struct ShellTabRow: View {
-    let tab: ShellTab
-    let isSelected: Bool
-
-    var body: some View {
-        HStack(spacing: 10) {
-            Image(systemName: iconName)
-                .font(.system(size: 11, weight: .semibold))
-                .foregroundStyle(isSelected ? ShellPalette.accent : ShellPalette.mutedInk)
-                .frame(width: 14)
-
-            VStack(alignment: .leading, spacing: 2) {
-                Text(tab.title ?? tab.kind.rawValue.capitalized)
-                    .font(.system(size: 13, weight: .semibold))
-                    .foregroundStyle(ShellPalette.ink)
-                    .lineLimit(1)
-                Text(tab.kind.rawValue.capitalized)
-                    .font(.system(size: 11, weight: .medium))
-                    .foregroundStyle(ShellPalette.mutedInk)
-            }
-
-            Spacer(minLength: 0)
-        }
-        .padding(.horizontal, 12)
-        .padding(.vertical, 10)
-        .background(
-            RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                .fill(isSelected ? ShellPalette.sidebarCard : Color.clear)
-        )
-        .overlay {
-            RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                .stroke(isSelected ? ShellPalette.line.opacity(0.65) : Color.clear, lineWidth: 1)
-        }
-    }
-
-    private var iconName: String {
-        switch tab.kind {
-        case .terminal:
-            return "terminal"
-        case .log:
-            return "doc.text.magnifyingglass"
-        case .scratch:
-            return "note.text"
-        }
-    }
-}
-
-private struct ShellAttentionRow: View {
-    let item: ShellAttentionItem
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 4) {
-            HStack {
-                Text(item.title)
-                    .font(.system(size: 12, weight: .semibold))
-                    .foregroundStyle(ShellPalette.ink)
-                Spacer(minLength: 8)
-                Text(item.attention.rawValue.replacingOccurrences(of: "_", with: " "))
-                    .font(.system(size: 10, weight: .semibold))
-                    .textCase(.uppercase)
-                    .foregroundStyle(item.attention == .awaitingUser ? ShellPalette.attention : ShellPalette.accent)
-            }
-            Text(item.summary)
-                .font(.system(size: 11, weight: .medium))
-                .foregroundStyle(ShellPalette.mutedInk)
-                .frame(maxWidth: .infinity, alignment: .leading)
-                .lineLimit(2)
-        }
-        .padding(.horizontal, 12)
-        .padding(.vertical, 10)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(
-            RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                .fill(ShellPalette.panel)
-        )
-    }
-}
-
-private struct ShellEmptyStateRow: View {
-    let title: String
-    let detail: String
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            Text(title)
-                .font(.system(size: 13, weight: .semibold))
-            Text(detail)
-                .font(.system(size: 12, weight: .medium))
-                .foregroundStyle(ShellPalette.mutedInk)
-        }
-        .padding(.horizontal, 12)
-        .padding(.vertical, 10)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(
-            RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                .fill(ShellPalette.panelSoft)
-        )
-    }
-}
-
-private enum ShellCommandTabAction: String, CaseIterable, Identifiable {
-    case newSpace
-    case newAlanSpace
-    case openTab
-    case openAlanTab
-    case jumpToAttention
-    case focusBestPane
-    case splitRight
-    case splitDown
-    case splitLeft
-    case splitUp
-    case focusLeft
-    case focusRight
-    case focusUp
-    case focusDown
-    case equalizeSplits
-    case liftPane
-    case closePane
-    case closeTab
-    case copySnapshot
-
-    var id: String { rawValue }
-
-    var title: String {
-        switch self {
-        case .newSpace:
-            return "Create Space"
-        case .newAlanSpace:
-            return "Create Space with Alan"
-        case .openTab:
-            return "Open New Tab"
-        case .openAlanTab:
-            return "Open In Alan"
-        case .jumpToAttention:
-            return "Jump To Attention"
-        case .focusBestPane:
-            return "Focus Best Routing Pane"
-        case .splitRight:
-            return "Split Pane Right"
-        case .splitDown:
-            return "Split Pane Down"
-        case .splitLeft:
-            return "Split Pane Left"
-        case .splitUp:
-            return "Split Pane Up"
-        case .focusLeft:
-            return "Focus Pane Left"
-        case .focusRight:
-            return "Focus Pane Right"
-        case .focusUp:
-            return "Focus Pane Up"
-        case .focusDown:
-            return "Focus Pane Down"
-        case .equalizeSplits:
-            return "Equalize Splits"
-        case .liftPane:
-            return "Lift Focused Pane To Tab"
-        case .closePane:
-            return "Close Focused Pane"
-        case .closeTab:
-            return "Close Current Tab"
-        case .copySnapshot:
-            return "Copy Shell Snapshot"
-        }
-    }
-
-    var detail: String {
-        switch self {
-        case .newSpace:
-            return "Start a fresh space with a plain login shell."
-        case .newAlanSpace:
-            return "Start a fresh space that opens directly into Alan."
-        case .openTab:
-            return "Open another tab inside the current space."
-        case .openAlanTab:
-            return "Open another tab that boots directly into Alan."
-        case .jumpToAttention:
-            return "Jump to the strongest pane that currently needs approval or attention."
-        case .focusBestPane:
-            return "Use shell routing signals to jump to the strongest candidate pane."
-        case .splitRight:
-            return "Create a side-by-side split to the right of the focused pane."
-        case .splitDown:
-            return "Create a stacked split beneath the focused pane."
-        case .splitLeft:
-            return "Create a side-by-side split to the left of the focused pane."
-        case .splitUp:
-            return "Create a stacked split above the focused pane."
-        case .focusLeft:
-            return "Move terminal focus to the pane on the left."
-        case .focusRight:
-            return "Move terminal focus to the pane on the right."
-        case .focusUp:
-            return "Move terminal focus to the pane above."
-        case .focusDown:
-            return "Move terminal focus to the pane below."
-        case .equalizeSplits:
-            return "Return the current tab's split dividers to equal ratios."
-        case .liftPane:
-            return "Move the focused pane into its own tab without losing shell identity."
-        case .closePane:
-            return "Close the focused pane and keep the remaining tab layout intact."
-        case .closeTab:
-            return "Close the current tab while preserving the rest of the space."
-        case .copySnapshot:
-            return "Copy the canonical shell JSON for debugging or agent context."
-        }
-    }
-
-    func matches(query: String) -> Bool {
-        query.isEmpty
-            || title.localizedCaseInsensitiveContains(query)
-            || detail.localizedCaseInsensitiveContains(query)
-            || keywordMatches(query: query)
-    }
-
-    private func keywordMatches(query: String) -> Bool {
-        let normalizedQuery = query.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-        guard !normalizedQuery.isEmpty else { return false }
-        return keywords.contains { normalizedQuery.contains($0) }
-    }
-
-    private var keywords: [String] {
-        switch self {
-        case .newSpace:
-            return ["new space", "fresh space", "workspace", "terminal space"]
-        case .newAlanSpace:
-            return ["new space with alan", "alan space", "agent space"]
-        case .openTab:
-            return ["open tab", "new tab", "open terminal tab"]
-        case .openAlanTab:
-            return ["open in alan", "alan tab", "new alan tab"]
-        case .jumpToAttention:
-            return ["jump to attention", "jump attention", "focus waiting pane", "approval", "waiting pane"]
-        case .focusBestPane:
-            return ["best pane", "route", "routing", "focus best", "jump pane"]
-        case .splitRight:
-            return ["split right", "split vertical", "side by side"]
-        case .splitDown:
-            return ["split down", "split horizontal", "split below", "stack split"]
-        case .splitLeft:
-            return ["split left"]
-        case .splitUp:
-            return ["split up", "split above"]
-        case .focusLeft:
-            return ["focus left", "pane left"]
-        case .focusRight:
-            return ["focus right", "pane right"]
-        case .focusUp:
-            return ["focus up", "pane up"]
-        case .focusDown:
-            return ["focus down", "pane down"]
-        case .equalizeSplits:
-            return ["equalize", "balance splits", "reset split ratios"]
-        case .liftPane:
-            return ["lift pane", "move pane", "extract pane"]
-        case .closePane:
-            return ["close pane", "remove pane"]
-        case .closeTab:
-            return ["close tab", "close tab"]
-        case .copySnapshot:
-            return ["copy snapshot", "copy json", "debug snapshot"]
-        }
-    }
-}
-
-private struct ShellCommandTabIntent: Identifiable {
-    enum Route {
-        case action(ShellCommandTabAction)
-        case attention(ShellAttentionItem)
-        case candidate(AlanShellRoutingCandidate)
-    }
-
-    let title: String
-    let detail: String
-    let accent: Color
-    let route: Route
-
-    var id: String { title }
-}
-
-private struct ShellCommandTabView: View {
-    @ObservedObject var host: ShellHostController
-    @Binding var isPresented: Bool
-    @State private var query = ""
-    @FocusState private var isQueryFocused: Bool
-    @StateObject private var voiceController = ShellVoiceCommandController()
-
-    private var matchingActions: [ShellCommandTabAction] {
-        let allActions = ShellCommandTabAction.allCases.filter { $0.matches(query: query) }
-
-        guard query.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty else {
-            return allActions
-        }
-
-        let defaultActions: [ShellCommandTabAction] = [
-            .openTab,
-            .openAlanTab,
-            .splitRight,
-            .splitDown,
-            .equalizeSplits,
-            .jumpToAttention,
-            .newSpace,
-        ]
-
-        return defaultActions.filter { allActions.contains($0) }
-    }
-
-    private var matchingAttention: [ShellAttentionItem] {
-        let visibleItems = host.attentionItems.filter {
-            $0.attention == .awaitingUser || $0.attention == .notable
-        }
-
-        guard !query.isEmpty else { return Array(visibleItems.prefix(2)) }
-        return visibleItems.filter {
-            $0.title.localizedCaseInsensitiveContains(query)
-                || $0.summary.localizedCaseInsensitiveContains(query)
-        }
-    }
-
-    private var matchingRoutingCandidates: [AlanShellRoutingCandidate] {
-        let candidates = host.routingCandidates
-        guard !query.isEmpty else { return Array(candidates.prefix(2)) }
-
-        let normalized = query.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-        return candidates.filter { candidate in
-            guard let pane = host.shellState.panes.first(where: { $0.paneID == candidate.paneID }) else {
-                return candidate.paneID.localizedCaseInsensitiveContains(query)
-            }
-
-            return candidate.paneID.localizedCaseInsensitiveContains(query)
-                || (pane.viewport?.title?.localizedCaseInsensitiveContains(query) ?? false)
-                || (pane.viewport?.summary?.localizedCaseInsensitiveContains(query) ?? false)
-                || (pane.process?.program.localizedCaseInsensitiveContains(query) ?? false)
-                || candidate.reasons.contains { $0.lowercased().contains(normalized) }
-        }
-    }
-
-    private var primaryIntent: ShellCommandTabIntent? {
-        let normalized = query.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-        guard !normalized.isEmpty else { return nil }
-
-        if normalized.contains("route")
-            || normalized.contains("best pane")
-            || normalized.contains("focus best")
-        {
-            if let candidate = matchingRoutingCandidates.first {
-                return ShellCommandTabIntent(
-                    title: "Focus \(routingTitle(for: candidate))",
-                    detail: routingDetail(for: candidate),
-                    accent: ShellPalette.accent,
-                    route: .candidate(candidate)
-                )
-            }
-        }
-
-        if normalized.contains("attention") || normalized.contains("waiting") || normalized.contains("jump") {
-            if let firstAttention = matchingAttention.first {
-                return ShellCommandTabIntent(
-                    title: "Jump To \(firstAttention.title)",
-                    detail: "Focus the pane that currently needs attention first.",
-                    accent: firstAttention.attention == .awaitingUser ? ShellPalette.accent : ShellPalette.ink,
-                    route: .attention(firstAttention)
-                )
-            }
-        }
-
-        if let action = matchingActions.first {
-            return ShellCommandTabIntent(
-                title: title(for: action),
-                detail: detail(for: action),
-                accent: ShellPalette.accent,
-                route: .action(action)
-            )
-        }
-
-        return nil
-    }
-
-    var body: some View {
-        ScrollView(showsIndicators: false) {
-            VStack(alignment: .leading, spacing: 14) {
-                HStack(alignment: .center, spacing: 12) {
-                    Image(systemName: "magnifyingglass")
-                        .font(.system(size: 14, weight: .semibold))
-                        .foregroundStyle(ShellPalette.accent)
-
-                    TextField(
-                        "",
-                        text: $query,
-                        prompt: Text("Go to or Command...")
-                            .foregroundStyle(ShellPalette.mutedInk.opacity(0.9))
-                    )
-                        .textFieldStyle(.plain)
-                        .font(.system(size: 16, weight: .medium))
-                        .foregroundStyle(ShellPalette.ink)
-                        .focused($isQueryFocused)
-                        .onSubmit {
-                            executePrimaryIntent()
-                        }
-
-                    Text("⌘K")
-                        .font(.system(size: 10, weight: .semibold, design: .monospaced))
-                        .foregroundStyle(ShellPalette.mutedInk.opacity(0.9))
-                        .padding(.horizontal, 7)
-                        .padding(.vertical, 4)
-                        .background(
-                            RoundedRectangle(cornerRadius: ShellRadii.control, style: .continuous)
-                                .fill(ShellPalette.canvas.opacity(0.95))
-                        )
-
-                    Button {
-                        voiceController.toggleListening { recognizedCommand in
-                            query = recognizedCommand
-                            executePrimaryIntent()
-                        }
-                    } label: {
-                        Image(systemName: voiceController.isListening ? "mic.fill" : "mic")
-                            .font(.system(size: 12, weight: .semibold))
-                            .foregroundStyle(voiceController.isListening ? ShellPalette.accent : ShellPalette.mutedInk)
-                            .frame(width: 26, height: 26)
-                            .background(
-                                RoundedRectangle(cornerRadius: ShellRadii.control, style: .continuous)
-                                    .fill(ShellPalette.canvas.opacity(0.92))
-                            )
-                    }
-                    .buttonStyle(.plain)
-
-                    Button {
-                        isPresented = false
-                    } label: {
-                        Image(systemName: "xmark")
-                            .font(.system(size: 11, weight: .bold))
-                            .foregroundStyle(ShellPalette.mutedInk)
-                            .frame(width: 26, height: 26)
-                            .background(
-                                RoundedRectangle(cornerRadius: ShellRadii.control, style: .continuous)
-                                    .fill(ShellPalette.canvas.opacity(0.92))
-                            )
-                    }
-                    .buttonStyle(.plain)
-                }
-                .padding(.horizontal, 14)
-                .padding(.vertical, 12)
-                .background(
-                    RoundedRectangle(cornerRadius: ShellRadii.surface, style: .continuous)
-                        .fill(Color.white.opacity(0.9))
-                )
-                .overlay {
-                    RoundedRectangle(cornerRadius: ShellRadii.surface, style: .continuous)
-                        .stroke(ShellPalette.line.opacity(0.32), lineWidth: 1)
-                }
-                .shadow(color: Color.black.opacity(0.035), radius: 10, y: 4)
-
-                if voiceController.isListening || primaryIntent != nil {
-                    VStack(alignment: .leading, spacing: 12) {
-                        if voiceController.isListening {
-                            HStack(spacing: 8) {
-                                Circle()
-                                    .fill(ShellPalette.attention)
-                                    .frame(width: 8, height: 8)
-                                Text("Listening for shell actions")
-                                    .font(.system(size: 12, weight: .medium))
-                                    .foregroundStyle(ShellPalette.mutedInk)
-                            }
-                        }
-
-                        if let primaryIntent {
-                            VStack(alignment: .leading, spacing: 10) {
-                                sectionLabel("Best match")
-                                Button {
-                                    execute(primaryIntent.route)
-                                } label: {
-                                    ShellCommandRow(
-                                        title: primaryIntent.title,
-                                        detail: primaryIntent.detail,
-                                        accent: primaryIntent.accent
-                                    )
-                                }
-                                .buttonStyle(.plain)
-                            }
-                        }
-                    }
-                }
-
-                VStack(alignment: .leading, spacing: 10) {
-                    sectionLabel("Actions")
-                    VStack(spacing: 8) {
-                        ForEach(matchingActions) { action in
-                            Button {
-                                perform(action)
-                            } label: {
-                                ShellCommandRow(
-                                    title: title(for: action),
-                                    detail: detail(for: action),
-                                    accent: ShellPalette.accent
-                                )
-                            }
-                            .buttonStyle(.plain)
-                        }
-                    }
-                }
-
-                if !matchingRoutingCandidates.isEmpty && !query.isEmpty {
-                    VStack(alignment: .leading, spacing: 10) {
-                        sectionLabel("Routing")
-                        VStack(spacing: 8) {
-                            ForEach(matchingRoutingCandidates) { candidate in
-                                Button {
-                                    execute(.candidate(candidate))
-                                } label: {
-                                    ShellCommandRow(
-                                        title: "Focus \(routingTitle(for: candidate))",
-                                        detail: routingDetail(for: candidate),
-                                        accent: ShellPalette.accent
-                                    )
-                                }
-                                .buttonStyle(.plain)
-                            }
-                        }
-                    }
-                }
-
-                if !matchingAttention.isEmpty {
-                    VStack(alignment: .leading, spacing: 10) {
-                        sectionLabel("Attention")
-                        VStack(spacing: 8) {
-                            ForEach(matchingAttention) { item in
-                                Button {
-                                    host.focusAttentionItem(item)
-                                    isPresented = false
-                                } label: {
-                                    ShellCommandRow(
-                                        title: shellNormalizedTitle(item.title) ?? item.title,
-                                        detail: shellUserFacingSummary(item.summary) ?? "Needs attention",
-                                        accent: item.attention == .awaitingUser ? ShellPalette.attention : ShellPalette.ink
-                                    )
-                                }
-                                .buttonStyle(.plain)
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        .padding(16)
-        .frame(width: 478, height: 568)
-        .background(
-            ZStack {
-                RoundedRectangle(cornerRadius: ShellRadii.overlay, style: .continuous)
-                    .fill(.ultraThinMaterial)
-                RoundedRectangle(cornerRadius: ShellRadii.overlay, style: .continuous)
-                    .fill(ShellPalette.window.opacity(0.9))
-            }
-        )
-        .overlay {
-            RoundedRectangle(cornerRadius: ShellRadii.overlay, style: .continuous)
-                .stroke(ShellPalette.line.opacity(0.42), lineWidth: 1)
-        }
-        .shadow(color: Color.black.opacity(0.12), radius: 26, y: 16)
-        .onAppear {
-            isQueryFocused = true
-        }
-        .onDisappear {
-            voiceController.stopListening()
-        }
-        .onExitCommand {
-            voiceController.stopListening()
-            isPresented = false
-        }
-    }
-
-    private func perform(_ action: ShellCommandTabAction) {
-        switch action {
-        case .newSpace:
-            _ = host.createTerminalSpace()
-        case .newAlanSpace:
-            _ = host.createAlanSpace()
-        case .openTab:
-            host.performShellWorkspaceCommand(.newTerminalTab)
-        case .openAlanTab:
-            host.performShellWorkspaceCommand(.newAlanTab)
-        case .jumpToAttention:
-            if let firstAttention = host.attentionItems.first {
-                host.focusAttentionItem(firstAttention)
-            }
-        case .focusBestPane:
-            _ = host.focusTopRoutingCandidate()
-        case .splitRight:
-            host.performShellWorkspaceCommand(ShellWorkspaceCommand.splitRight)
-        case .splitDown:
-            host.performShellWorkspaceCommand(ShellWorkspaceCommand.splitDown)
-        case .splitLeft:
-            host.performShellWorkspaceCommand(ShellWorkspaceCommand.splitLeft)
-        case .splitUp:
-            host.performShellWorkspaceCommand(ShellWorkspaceCommand.splitUp)
-        case .focusLeft:
-            host.performShellWorkspaceCommand(ShellWorkspaceCommand.focusLeft)
-        case .focusRight:
-            host.performShellWorkspaceCommand(ShellWorkspaceCommand.focusRight)
-        case .focusUp:
-            host.performShellWorkspaceCommand(ShellWorkspaceCommand.focusUp)
-        case .focusDown:
-            host.performShellWorkspaceCommand(ShellWorkspaceCommand.focusDown)
-        case .equalizeSplits:
-            host.performShellWorkspaceCommand(ShellWorkspaceCommand.equalizeSplits)
-        case .liftPane:
-            _ = host.liftSelectedPaneToTab()
-        case .closePane:
-            host.performShellWorkspaceCommand(.closePane)
-        case .closeTab:
-            host.performShellWorkspaceCommand(.closeTab)
-        case .copySnapshot:
-            host.copySnapshotJSON()
-        }
-        isPresented = false
-    }
-
-    private func execute(_ route: ShellCommandTabIntent.Route) {
-        switch route {
-        case let .action(action):
-            perform(action)
-        case let .attention(item):
-            host.focusAttentionItem(item)
-            isPresented = false
-        case let .candidate(candidate):
-            host.focus(paneID: candidate.paneID)
-            isPresented = false
-        }
-    }
-
-    private func executePrimaryIntent() {
-        guard let primaryIntent else { return }
-        execute(primaryIntent.route)
-    }
-
-    private func routingDetail(for candidate: AlanShellRoutingCandidate) -> String {
-        let title = routingTitle(for: candidate)
-        let reasons = candidate.reasons.prefix(3).joined(separator: " • ")
-        let detail = reasons.isEmpty ? "score \(Int(candidate.score * 100))" : reasons
-        return "\(title) • \(detail)"
-    }
-
-    private func routingTitle(for candidate: AlanShellRoutingCandidate) -> String {
-        guard let pane = host.shellState.panes.first(where: { $0.paneID == candidate.paneID }) else {
-            return "Terminal Pane"
-        }
-
-        return shellDisplayTitle(
-            rawTitle: pane.viewport?.title,
-            workingDirectoryName: pane.context?.workingDirectoryName,
-            cwd: pane.cwd,
-            program: pane.process?.program,
-            launchTarget: pane.resolvedLaunchTarget,
-            fallback: pane.viewport?.summary
-        )
-    }
-
-    private func title(for action: ShellCommandTabAction) -> String {
-        action.title
-    }
-
-    private func detail(for action: ShellCommandTabAction) -> String {
-        action.detail
-    }
-
-    private func sectionLabel(_ value: String) -> some View {
-        Text(value)
-            .font(.system(size: 10, weight: .semibold, design: .rounded))
-            .textCase(.uppercase)
-            .foregroundStyle(ShellPalette.mutedInk)
-    }
-}
-
-@MainActor
-private final class ShellVoiceCommandController: NSObject, ObservableObject, NSSpeechRecognizerDelegate {
-    @Published private(set) var isListening = false
-
-    private let recognizer = NSSpeechRecognizer()
-    private var recognitionHandler: ((String) -> Void)?
-
-    override init() {
-        super.init()
-        recognizer?.delegate = self
-        recognizer?.listensInForegroundOnly = false
-        recognizer?.blocksOtherRecognizers = false
-        recognizer?.commands = [
-            "new space",
-            "new alan space",
-            "open tab",
-            "open in alan",
-            "focus best pane",
-            "route to best pane",
-            "split right",
-            "split down",
-            "split left",
-            "split up",
-            "focus left",
-            "focus right",
-            "focus up",
-            "focus down",
-            "equalize splits",
-            "lift pane",
-            "close pane",
-            "close tab",
-            "jump to attention",
-            "focus waiting pane",
-            "copy snapshot",
-        ]
-    }
-
-    func toggleListening(handler: @escaping (String) -> Void) {
-        isListening ? stopListening() : startListening(handler: handler)
-    }
-
-    func startListening(handler: @escaping (String) -> Void) {
-        recognitionHandler = handler
-        recognizer?.startListening()
-        isListening = recognizer != nil
-    }
-
-    func stopListening() {
-        recognizer?.stopListening()
-        isListening = false
-        recognitionHandler = nil
-    }
-
-    func speechRecognizer(_ sender: NSSpeechRecognizer, didRecognizeCommand command: String) {
-        recognitionHandler?(command)
-    }
-}
-
-private struct ShellCommandRow: View {
-    @Environment(\.accessibilityReduceMotion) private var reduceMotion
-    @State private var isHovered = false
-    let title: String
-    let detail: String
-    let accent: Color
-
-    var body: some View {
-        HStack(alignment: .center, spacing: 12) {
-            Circle()
-                .fill(accent)
-                .frame(width: 7, height: 7)
-
-            VStack(alignment: .leading, spacing: 3) {
-                Text(title)
-                    .font(.system(size: 13, weight: .semibold))
-                    .foregroundStyle(ShellPalette.ink)
-                Text(detail)
-                    .font(.system(size: 11, weight: .medium))
-                    .foregroundStyle(ShellPalette.mutedInk)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .lineLimit(2)
-            }
-
-            Spacer(minLength: 0)
-        }
-        .padding(.horizontal, 14)
-        .padding(.vertical, 10)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(
-            RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                .fill(Color.white.opacity(0.84))
-        )
-        .overlay {
-            RoundedRectangle(cornerRadius: ShellRadii.row, style: .continuous)
-                .stroke(ShellPalette.line.opacity(isHovered ? 0.24 : 0.12), lineWidth: 1)
-        }
-        .scaleEffect(isHovered ? 1.004 : 1)
-        .shadow(color: isHovered ? Color.black.opacity(0.022) : .clear, radius: 6, y: 3)
-        .animation(reduceMotion ? nil : .easeOut(duration: 0.16), value: isHovered)
-        .onHover { isHovered = $0 }
-    }
-}
-
-func shellTabDisplayTitle(
-    tab: ShellTab?,
-    panes: [ShellPane],
-    fallback: String
-) -> String {
-    guard let tab else { return fallback }
-
-    if let title = shellNormalizedTitle(tab.title) {
-        return title
-    }
-
-    let preferredPane = panes.first(where: { $0.resolvedLaunchTarget == .alan }) ?? panes.first
-
-    return shellDisplayTitle(
-        rawTitle: preferredPane?.viewport?.title,
-        workingDirectoryName: preferredPane?.context?.workingDirectoryName,
-        cwd: preferredPane?.cwd,
-        program: preferredPane?.process?.program,
-        launchTarget: preferredPane?.resolvedLaunchTarget ?? .shell,
-        fallback: fallback
-    )
 }
 #endif
